# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VetHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics/{clinicId}:
    patch:
      operationId: updateClinic
      tags: [clinics]
      summary: Update clinic details
      parameters:
        - name: clinicId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicUpdate"
      responses:
        "200":
          description: Updated clinic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"

  # ── Staff ─────────────────────────────────────────────────────────────────
  /clinics/{clinicId}/staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff members of a clinic
      parameters:
        - name: clinicId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: inviteStaff
      tags: [staff]
      summary: Invite a veterinarian to the clinic by email or phone
      parameters:
        - name: clinicId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInviteInput"
      responses:
        "201":
          description: Staff invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /clinics/{clinicId}/staff/{staffId}:
    delete:
      operationId: removeStaff
      tags: [staff]
      summary: Remove a staff member from the clinic
      parameters:
        - name: clinicId
          in: path
          required: true
          schema: { type: integer }
        - name: staffId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed

  # ── Species ───────────────────────────────────────────────────────────────
...
  /daily-reports/{reportId}:
    get:
      operationId: getDailyReport
      tags: [daily-reports]
      summary: Get a daily report
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Daily report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
    patch:
      operationId: updateDailyReport
      tags: [daily-reports]
      summary: Update a daily report
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyReportUpdate"
      responses:
        "200":
          description: Updated daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"

  # ── Products (POS) ────────────────────────────────────────────────────────
  /clinics/{clinicId}/products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products for a clinic
      parameters:
        - name: clinicId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Add a product to the clinic catalog
      parameters:
        - name: clinicId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /clinics/{clinicId}/reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get daily/monthly/yearly revenue and visit summary
      parameters:
        - name: clinicId
          in: path
          required: true
          schema: { type: integer }
        - name: period
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ReportPeriod"
        - name: date
          in: query
          required: true
          description: ISO date string (YYYY-MM-DD) for reference point
          schema: { type: string }
      responses:
        "200":
          description: Report summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /clinics/{clinicId}/reports/visits:
    get:
      operationId: getVisitStats
      tags: [reports]
      summary: Get visit count statistics
      parameters:
        - name: clinicId
          in: path
          required: true
          schema: { type: integer }
        - name: period
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ReportPeriod"
        - name: date
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Visit stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitStats"
  /clinics/{clinicId}/visits/active:
    get:
      operationId: listActiveVisits
      tags: [visits]
      summary: Get all currently active/inpatient visits for a clinic
      parameters:
        - name: clinicId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Active visits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VisitWithPet"
components:
  schemas:
...
    ReportPeriod:
      type: string
      enum: [daily, weekly, monthly, quarterly, yearly]
...
    Clinic:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        address: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        ownerId: { type: integer }
        createdAt: { type: string }
      required: [id, name, ownerId, createdAt]
...
    ClinicUpdate:
      type: object
      properties:
        name: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
    StaffMember:
      type: object
      properties:
        id: { type: integer }
        clinicId: { type: integer }
        userId: { type: integer }
        name: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        role: { type: string }
        status: { type: string }
        createdAt: { type: string }
      required: [id, clinicId, userId, role, status, createdAt]
    StaffInviteInput:
      type: object
      properties:
        email: { type: string }
        phone: { type: string }
        name: { type: string }
      required: [name]
...
    VisitWithPet:
      type: object
      properties:
        id: { type: integer }
        petId: { type: integer }
        petName: { type: string }
        speciesName: { type: ["string", "null"] }
        petStatus: { type: string }
        ownerName: { type: ["string", "null"] }
        ownerPhone: { type: ["string", "null"] }
        clinicId: { type: integer }
        type: { type: string }
        status: { type: string }
        visitDate: { type: string }
        totalCost: { type: number }
        latestReport: { type: ["string", "null"] }
      required:
        [id, petId, petName, clinicId, type, status, visitDate, totalCost]
...
    DailyReport:
      type: object
      properties:
        id: { type: integer }
        visitId: { type: integer }
        reportDate: { type: string }
        condition: { type: ["string", "null"] }
        treatment: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        cost: { type: number }
        vetName: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, visitId, reportDate, cost, createdAt]
...
    DailyReportUpdate:
      type: object
      properties:
        condition: { type: string }
        treatment: { type: string }
        notes: { type: string }
        cost: { type: number }
    Product:
      type: object
      properties:
        id: { type: integer }
        clinicId: { type: integer }
        name: { type: string }
        category: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        price: { type: number }
        stock: { type: ["integer", "null"] }
        unit: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
      required: [id, clinicId, name, price, isActive, createdAt]
    ProductInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        description: { type: string }
        price: { type: number }
        stock: { type: integer }
        unit: { type: string }
      required: [name, price]
...
    ReportSummary:
      type: object
      properties:
        period: { type: string }
        date: { type: string }
        totalRevenue: { type: number }
        totalVisits: { type: integer }
        inpatientVisits: { type: integer }
        outpatientVisits: { type: integer }
        averageRevenuePerVisit: { type: number }
        diedCount:
          {
            type: integer,
            description: "Inpatient visits where pet passed away",
          }
        survivedCount:
          {
            type: integer,
            description: "Completed inpatient visits where pet survived",
          }
        earlyDischargeCount:
          {
            type: integer,
            description: "Visits cancelled before completion (taken early by owner)",
          }
        topServices:
          type: array
          items:
            $ref: "#/components/schemas/TopItem"
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/TopItem"
      required:
        [
          period,
          date,
          totalRevenue,
          totalVisits,
          inpatientVisits,
          outpatientVisits,
          averageRevenuePerVisit,
          diedCount,
          survivedCount,
          earlyDischargeCount,
          topServices,
          topProducts,
        ]
    TopItem:
      type: object
      properties:
        name: { type: string }
        count: { type: integer }
        revenue: { type: number }
      required: [name, count, revenue]
    VisitStats:
      type: object
      properties:
        period: { type: string }
        labels:
          type: array
          items: { type: string }
        visitCounts:
          type: array
          items: { type: integer }
        revenues:
          type: array
          items: { type: number }
      required: [period, labels, visitCounts, revenues]